Extract search modal, results, and their tests from TopBar
